# apps/monolith/spec/swagger.json
{
	"components": {
...
		"schemas": {
			"ApplicationStatus": {
				"enum": [
					"WAITING",
					"ACCEPTED",
					"DECLINED"
				],
				"type": "string"
			},
			"Partial_Pick_User.id-or-username-or-avatar__": {
				"properties": {
					"id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"avatar": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"UserHeader": {
				"$ref": "#/components/schemas/Partial_Pick_User.id-or-username-or-avatar__"
			},
			"Partial_Pick_Table.id-or-owner-or-title-or-banner__": {
				"properties": {
					"id": {
						"type": "string"
					},
					"owner": {
						"$ref": "#/components/schemas/UserHeader"
					},
					"title": {
						"type": "string"
					},
					"banner": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"TableHeader": {
				"$ref": "#/components/schemas/Partial_Pick_Table.id-or-owner-or-title-or-banner__"
			},
			"Application": {
				"properties": {
					"id": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/ApplicationStatus"
					},
					"applicant": {
						"$ref": "#/components/schemas/UserHeader"
					},
					"table": {
						"$ref": "#/components/schemas/TableHeader"
					},
					"message": {
						"type": "string"
					},
					"creationDate": {
						"type": "string",
						"format": "date-time"
					},
					"lastUpdateDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"status",
					"applicant",
					"table",
					"message",
					"creationDate",
					"lastUpdateDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_ApplicationCreateBody.applicant-or-table-or-message_": {
				"properties": {
					"applicant": {
						"$ref": "#/components/schemas/UserHeader"
					},
					"table": {
						"$ref": "#/components/schemas/TableHeader"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"applicant",
					"table",
					"message"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"ApplicationCreateBodyRequest": {
				"$ref": "#/components/schemas/Pick_ApplicationCreateBody.applicant-or-table-or-message_"
			},
			"AuthUserRegisterBody": {
				"properties": {
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial__email-string--password-string__": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"AuthUserUpdateBody": {
				"$ref": "#/components/schemas/Partial__email-string--password-string__"
			},
			"FlairTypes": {
				"enum": [
					"SYSTEM",
					"LANGUAGES",
					"RATINGS",
					"VACANCIES",
					"GENRES",
					"CUSTOM"
				],
				"type": "string"
			},
			"Record_SupportedLanguages.string_": {
				"properties": {
					"en": {
						"type": "string"
					},
					"pt": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"Flair": {
				"properties": {
					"id": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/FlairTypes"
					},
					"name": {
						"type": "string"
					},
					"labels": {
						"$ref": "#/components/schemas/Record_SupportedLanguages.string_"
					},
					"numberOfUses": {
						"type": "number",
						"format": "double"
					},
					"creationDate": {
						"type": "string",
						"format": "date-time"
					},
					"lastUpdateDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"type",
					"name",
					"labels",
					"numberOfUses",
					"creationDate",
					"lastUpdateDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_FlairCreationBody.type-or-name-or-labels_": {
				"properties": {
					"type": {
						"$ref": "#/components/schemas/FlairTypes"
					},
					"name": {
						"type": "string"
					},
					"labels": {
						"$ref": "#/components/schemas/Record_SupportedLanguages.string_"
					}
				},
				"required": [
					"type",
					"name",
					"labels"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"FlairCreationBodyRequest": {
				"$ref": "#/components/schemas/Pick_FlairCreationBody.type-or-name-or-labels_"
			},
			"Pick_FlairUpdateBody.type-or-name-or-labels_": {
				"properties": {
					"type": {
						"$ref": "#/components/schemas/FlairTypes"
					},
					"name": {
						"type": "string"
					},
					"labels": {
						"$ref": "#/components/schemas/Record_SupportedLanguages.string_"
					}
				},
				"required": [
					"type",
					"name",
					"labels"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"FlairUpdateBodyRequest": {
				"$ref": "#/components/schemas/Pick_FlairUpdateBody.type-or-name-or-labels_"
			},
			"NotificationType": {
				"enum": [
					"application",
					"report",
					"system"
				],
				"type": "string"
			},
			"NotificationContent": {
				"enum": [
					"applied-to-your-table",
					"application-accepted",
					"application-declined"
				],
				"type": "string"
			},
			"Partial_Pick_Application.id-or-applicant-or-table__": {
				"properties": {
					"id": {
						"type": "string"
					},
					"applicant": {
						"$ref": "#/components/schemas/UserHeader"
					},
					"table": {
						"$ref": "#/components/schemas/TableHeader"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"ApplicationHeader": {
				"$ref": "#/components/schemas/Partial_Pick_Application.id-or-applicant-or-table__"
			},
			"NotificationData": {
				"$ref": "#/components/schemas/ApplicationHeader"
			},
			"Notification": {
				"properties": {
					"id": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/NotificationType"
					},
					"content": {
						"$ref": "#/components/schemas/NotificationContent"
					},
					"read": {
						"type": "boolean"
					},
					"creationDate": {
						"type": "string",
						"format": "date-time"
					},
					"lastUpdateDate": {
						"type": "string",
						"format": "date-time"
					},
					"data": {
						"$ref": "#/components/schemas/NotificationData"
					}
				},
				"required": [
					"id",
					"userId",
					"type",
					"content",
					"read",
					"creationDate",
					"lastUpdateDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Table": {
				"properties": {
					"id": {
						"type": "string"
					},
					"owner": {
						"$ref": "#/components/schemas/UserHeader"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"banner": {
						"type": "string"
					},
					"flairs": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"acceptMessage": {
						"type": "string"
					},
					"creationDate": {
						"type": "string",
						"format": "date-time"
					},
					"lastUpdateDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"owner",
					"title",
					"description",
					"banner",
					"flairs",
					"acceptMessage",
					"creationDate",
					"lastUpdateDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_TableCreateBody.title-or-description-or-banner-or-flairs-or-acceptMessage_": {
				"properties": {
					"title": {
						"type": "string"
					},
					"banner": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"flairs": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"acceptMessage": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"banner",
					"description",
					"flairs",
					"acceptMessage"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"TableCreateBodyRequest": {
				"$ref": "#/components/schemas/Pick_TableCreateBody.title-or-description-or-banner-or-flairs-or-acceptMessage_"
			},
			"Pick_TableUpdateBody.title-or-description-or-banner-or-flairs-or-acceptMessage_": {
				"properties": {
					"title": {
						"type": "string"
					},
					"banner": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"flairs": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"acceptMessage": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"banner",
					"description",
					"flairs",
					"acceptMessage"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"TableUpdateBodyRequest": {
				"$ref": "#/components/schemas/Pick_TableUpdateBody.title-or-description-or-banner-or-flairs-or-acceptMessage_"
			},
...
		},
		"securitySchemes": {
			"Bearer": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
	"info": {
		"title": "@rpg-together/source",
		"version": "0.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/applications": {
			"post": {
				"operationId": "CreateApplication",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Application"
								}
							}
						}
					}
				},
				"tags": [
					"Applications Service"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ApplicationCreateBodyRequest"
							}
						}
					}
				}
			}
		},
		"/applications/from-user/{userId}": {
			"get": {
				"operationId": "GetApplicationsFromUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Application"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Applications Service"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/applications/from-table/{tableId}": {
			"get": {
				"operationId": "GetApplicationsFromTable",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Application"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Applications Service"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tableId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/applications/from-user/{userId}/from-table/{tableId}": {
			"get": {
				"operationId": "GetApplicationFromUserAndTable",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Application"
								}
							}
						}
					}
				},
				"tags": [
					"Applications Service"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "tableId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/applications/{applicationId}": {
			"get": {
				"operationId": "GetApplication",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Application"
								}
							}
						}
					}
				},
				"tags": [
					"Applications Service"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "applicationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteApplication",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Applications Service"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "applicationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/applications/{applicationId}/accept": {
			"post": {
				"operationId": "AcceptApplication",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Applications Service"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "applicationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/applications/{applicationId}/decline": {
			"post": {
				"operationId": "DeclineApplication",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Applications Service"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "applicationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/auth/register/user": {
			"post": {
				"operationId": "UserRegister",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Authentication Service"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthUserRegisterBody"
							}
						}
					}
				}
			}
		},
		"/auth/register/admin": {
			"post": {
				"operationId": "AdminRegister",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Authentication Service"
				],
				"security": [
					{
						"Bearer": [
							"ADMIN"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthUserRegisterBody"
							}
						}
					}
				}
			}
		},
		"/auth/update": {
			"put": {
				"operationId": "UpdateAuthUser",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Authentication Service"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthUserUpdateBody"
							}
						}
					}
				}
			}
		},
		"/auth/delete": {
			"delete": {
				"operationId": "DeleteAuthUser",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Authentication Service"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": []
			}
		},
		"/flairs": {
			"post": {
				"operationId": "CreateFlair",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Flair"
								}
							}
						}
					}
				},
				"tags": [
					"Flairs Service"
				],
				"security": [
					{
						"Bearer": [
							"ADMIN"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/FlairCreationBodyRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllFlairs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Flair"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Flairs Service"
				],
				"security": [],
				"parameters": []
			}
		},
		"/flairs/{flairId}": {
			"get": {
				"operationId": "GetFlair",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Flair"
								}
							}
						}
					}
				},
				"tags": [
					"Flairs Service"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "flairId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateFlair",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Flair"
								}
							}
						}
					}
				},
				"tags": [
					"Flairs Service"
				],
				"security": [
					{
						"Bearer": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "flairId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/FlairUpdateBodyRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteFlair",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Flairs Service"
				],
				"security": [
					{
						"Bearer": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "flairId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/notifications/from-user/{userId}": {
			"get": {
				"operationId": "GetNotificationsFromUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Notification"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Notifications Service"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/notifications/{notificationId}": {
			"get": {
				"operationId": "GetNotification",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Notification"
								}
							}
						}
					}
				},
				"tags": [
					"Notifications Service"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "notificationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/notifications/{notificationId}/read": {
			"post": {
				"operationId": "ReadNotification",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Notifications Service"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "notificationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/tables": {
			"post": {
				"operationId": "CreateTable",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Table"
								}
							}
						}
					}
				},
				"tags": [
					"Tables Service"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TableCreateBodyRequest"
							}
						}
					}
				}
			}
		},
		"/tables/from-user/{userId}": {
			"get": {
				"operationId": "GetTablesFromUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Table"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Tables Service"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/tables/{tableId}": {
			"get": {
				"operationId": "GetTable",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Table"
								}
							}
						}
					}
				},
				"tags": [
					"Tables Service"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "tableId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateTable",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Table"
								}
							}
						}
					}
				},
				"tags": [
					"Tables Service"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tableId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TableUpdateBodyRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteTable",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Tables Service"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tableId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}